cleanup unnecessary code
