+++ packages/lesswrong/lib/components.ts
@@ -668,7 +668,7 @@ importComponent("RecommendationsAndCurated", () => require('../components/recomm
 importComponent("SpotlightHistory", () => require('../components/spotlights/SpotlightHistory'));
 importComponent("SpotlightItem", () => require('../components/spotlights/SpotlightItem'));
 importComponent("SpotlightEditorStyles", () => require('../components/spotlights/SpotlightEditorStyles'));
-
+importComponent("SpotlightStartOrContinueReading", () => require('../components/spotlights/SpotlightStartOrContinueReading'));
 importComponent("SpotlightsPage", () => require('../components/spotlights/SpotlightsPage'));
 importComponent("CurrentSpotlightItem", () => require('../components/spotlights/CurrentSpotlightItem'));
